revision - Sneha
is this what you were looking for? the reason that they show up as some space in between is because you have one table with one row with 4 columns next to each other. if you want them to be further spaced out from each other, you have to increase the width between the columns